Added main file with first functionalities
So far this script can read .onnx files and convert them into
dictionaries. There are two main functions that accomplish this:
get_model_dict() and get_model_params. The first one gets the main
structure of the model returning it as a dictionary. The second one
grabs inputs and outputs from the .onnx file and cand return them
both as a dictionary and as a list.

A third function can convert various dictionaries into onnx
GraphProtos that can be the converted to models and saved as .onnx
files.

After the models are converted to dictionaries they can be
manipulated and changed in any way a dictionary can be changed.